http://jira.sakaiproject.org/jira/browse/EVALSYS-394 Updated preview item to use the display number when previewing items in templates
git-svn-id: https://source.sakaiproject.org/contrib/evaluation/trunk@46261 fdecad78-55fc-0310-b1b2-d7d25cf747c9

+++ api/src/java/org/sakaiproject/evaluation/utils/TemplateItemUtils.java
@@ -315,7 +315,7 @@ public class TemplateItemUtils {
       }
 
       EvalTemplateItem templateItem = new EvalTemplateItem(item.getLastModified(), item.getOwner(), null, item,
-            new Integer(1), item.getCategory(), EvalConstants.HIERARCHY_LEVEL_TOP, EvalConstants.HIERARCHY_NODE_ID_NONE);
+            new Integer(0), item.getCategory(), EvalConstants.HIERARCHY_LEVEL_TOP, EvalConstants.HIERARCHY_NODE_ID_NONE);
 
       if (item.getUsesNA() == null) {
          templateItem.setUsesNA(Boolean.TRUE);
